add package-lock.json
it's good practice to add this to repos as it ensures that everyone (and
CI) use the same versions of dependencies.

in the upcoming commits you'll see formating changes that are due to a
newer version of prettier. with a lockfile those would not have happened
until someone intentionally updated that dependency.